feat(analysis): creates TPZEigenAnalysisBase
